# doc/apidoc/schema_swagger_json.json
{
  "swagger": "2.0",
  "info": {
    "title": "Finances API (params in:body)",
    "description": "\n<p>Financial management API</p>\n",
    "version": "1.0",
    "x-copyright": null
  },
  "basePath": "",
  "consumes": [
    "application/json"
  ],
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "get_v1_accounts",
        "summary": "List all user accounts",
        "parameters": [
          {
            "name": "Authentication",
            "in": "header",
            "required": true,
            "description": "User access token",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/account_response"
              }
            }
          }
        },
        "description": ""
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "post_v1_accounts",
        "summary": "Creates a account",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Account name",
                  "default": null
                },
                "description": {
                  "type": "string",
                  "description": "Account description",
                  "default": null
                },
                "account_type": {
                  "type": "string",
                  "enum": [
                    "checking",
                    "savings",
                    "other"
                  ],
                  "description": "Account Type",
                  "default": null
                },
                "initial_amount": {
                  "type": "string",
                  "description": "Account initial amount",
                  "default": null
                },
                "financial_institution_id": {
                  "type": "number",
                  "description": "Account related financial_institution id",
                  "default": null
                }
              },
              "additionalProperties": false,
              "required": [
                "account_type",
                "initial_amount",
                "financial_institution_id"
              ]
            }
          },
          {
            "name": "Authentication",
            "in": "header",
            "required": true,
            "description": "User access token",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "201": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "description": "Account id"
                },
                "name": {
                  "type": "string",
                  "description": "Account name"
                },
                "description": {
                  "type": "string",
                  "description": "Account description"
                },
                "financial_institution": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "description": "Financial Institution id"
                    },
                    "name": {
                      "type": "string",
                      "description": "Financial Institution name"
                    },
                    "logo_url": {
                      "type": "string",
                      "description": "Financial Institution logo url"
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "id",
                    "name",
                    "logo_url"
                  ]
                },
                "initial_amount": {
                  "type": "string",
                  "description": "Account initial amount"
                },
                "account_type": {
                  "type": "string",
                  "enum": [
                    "checking",
                    "savings",
                    "other"
                  ]
                }
              },
              "additionalProperties": false,
              "required": [
                "id",
                "name",
                "description",
                "financial_institution",
                "initial_amount",
                "account_type"
              ]
            }
          }
        },
        "description": ""
      }
    },
    "/v1/accounts/{id}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "get_v1_accounts_id",
        "summary": "Returns an account",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true,
            "description": "Transaction id"
          },
          {
            "name": "Authentication",
            "in": "header",
            "required": true,
            "description": "User access token",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "description": "Account id"
                },
                "name": {
                  "type": "string",
                  "description": "Account name"
                },
                "description": {
                  "type": "string",
                  "description": "Account description"
                },
                "financial_institution": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "description": "Financial Institution id"
                    },
                    "name": {
                      "type": "string",
                      "description": "Financial Institution name"
                    },
                    "logo_url": {
                      "type": "string",
                      "description": "Financial Institution logo url"
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "id",
                    "name",
                    "logo_url"
                  ]
                },
                "initial_amount": {
                  "type": "string",
                  "description": "Account initial amount"
                },
                "account_type": {
                  "type": "string",
                  "enum": [
                    "checking",
                    "savings",
                    "other"
                  ]
                }
              },
              "additionalProperties": false,
              "required": [
                "id",
                "name",
                "description",
                "financial_institution",
                "initial_amount",
                "account_type"
              ]
            }
          }
        },
        "description": ""
      },
      "put": {
        "tags": [
          "accounts"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "put_v1_accounts_id",
        "summary": "Updates a account",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true,
            "description": "Account id",
            "default": null
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Account name",
                  "default": null
                },
                "description": {
                  "type": "string",
                  "description": "Account description",
                  "default": null
                },
                "account_type": {
                  "type": "string",
                  "enum": [
                    "checking",
                    "savings",
                    "other"
                  ],
                  "description": "Account Type",
                  "default": null
                },
                "initial_amount": {
                  "type": "string",
                  "description": "Account initial amount",
                  "default": null
                },
                "financial_institution_id": {
                  "type": "number",
                  "description": "Account related financial_institution id",
                  "default": null
                }
              },
              "additionalProperties": false
            }
          },
          {
            "name": "Authentication",
            "in": "header",
            "required": true,
            "description": "User access token",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "description": "Account id"
                },
                "name": {
                  "type": "string",
                  "description": "Account name"
                },
                "description": {
                  "type": "string",
                  "description": "Account description"
                },
                "financial_institution": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "description": "Financial Institution id"
                    },
                    "name": {
                      "type": "string",
                      "description": "Financial Institution name"
                    },
                    "logo_url": {
                      "type": "string",
                      "description": "Financial Institution logo url"
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "id",
                    "name",
                    "logo_url"
                  ]
                },
                "initial_amount": {
                  "type": "string",
                  "description": "Account initial amount"
                },
                "account_type": {
                  "type": "string",
                  "enum": [
                    "checking",
                    "savings",
                    "other"
                  ]
                }
              },
              "additionalProperties": false,
              "required": [
                "id",
                "name",
                "description",
                "financial_institution",
                "initial_amount",
                "account_type"
              ]
            }
          }
        },
        "description": ""
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "consumes": [
          "application/json"
        ],
        "operationId": "delete_v1_accounts_id",
        "summary": "Delete a account",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true,
            "description": "Account id"
          },
          {
            "name": "Authentication",
            "in": "header",
            "required": true,
            "description": "User access token",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "204": {
            "description": "Successful response"
          }
        },
        "description": ""
      }
    },
...
  },
  "definitions": {
...
    "account_response": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "description": "Account id"
        },
        "name": {
          "type": "string",
          "description": "Account name"
        },
        "description": {
          "type": "string",
          "description": "Account description"
        },
        "financial_institution": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "description": "Financial Institution id"
            },
            "name": {
              "type": "string",
              "description": "Financial Institution name"
            },
            "logo_url": {
              "type": "string",
              "description": "Financial Institution logo url"
            }
          },
          "additionalProperties": false,
          "required": [
            "id",
            "name",
            "logo_url"
          ]
        },
        "initial_amount": {
          "type": "string",
          "description": "Account initial amount"
        },
        "account_type": {
          "type": "string",
          "enum": [
            "checking",
            "savings",
            "other"
          ]
        }
      },
      "additionalProperties": false,
      "required": [
        "id",
        "name",
        "description",
        "financial_institution",
        "initial_amount",
        "account_type"
      ]
    },
...
  },
...
  "host": "financesapi.herokuapp.com"
}